Update 5SENSORLINEFOLLOW.ino
Gripper function add
Modified all white line sensor mode. It checks if there is a corner and else go forward until it needs to turn. (we need three ultrasonic sensors)
Modified all black line sensor mode. It checks if it is a cross junction or a T junction. Same with the other type of junctions.
Action to follow the exit path if the three cylinders are found.